+++ src/chip.tsx
@@ -1,7 +1,7 @@
 'use client';
 import * as React from 'react';
 import classnames from 'classnames';
-import {useScreenSize, useTheme} from './hooks';
+import {useTheme} from './hooks';
 import Badge from './badge';
 import Box from './box';
 import {Text2} from './text';
@@ -63,11 +63,10 @@ const Chip: React.FC<ChipProps> = (props: ChipProps) => {
     const {texts, isDarkMode, textPresets} = useTheme();
 
     const overAlternative = useThemeVariant() === 'alternative';
-    const {isTabletOrSmaller} = useScreenSize();
 
-    const paddingLeft = Icon ? (isTabletOrSmaller ? 16 : 8) : isTabletOrSmaller ? 20 : 12;
-    const paddingRight = isTabletOrSmaller ? 20 : 12;
-    const paddingIcon = isTabletOrSmaller ? 16 : 8;
+    const paddingLeft = Icon ? ({mobile: 16, desktop: 8} as const) : ({mobile: 20, desktop: 12} as const);
+    const paddingRight = {mobile: 20, desktop: 12} as const;
+    const paddingIcon = {mobile: 16, desktop: 8} as const;
 
     const renderBadge = () => {
         if (!badge) {