+++ src/index.tsx
@@ -29,6 +29,7 @@ export {
     SuccessFeedbackScreen,
     SuccessFeedback,
 } from './feedback';
+export {LoadingScreen, BrandLoadingScreen} from './loading-screen';
 export {default as IconButton} from './icon-button';
 export {default as Popover} from './popover';
 export {default as FocusTrap} from './focus-trap';