+++ src/inline.css.ts
@@ -1,8 +1,38 @@
-import {style, createVar, globalStyle, styleVariants} from '@vanilla-extract/css';
+import {style, createVar, globalStyle, fallbackVar} from '@vanilla-extract/css';
+import * as mq from './media-queries.css';
 
 const space = createVar();
+const spaceMobile = createVar();
+const spaceTablet = createVar();
+const spaceDesktop = createVar();
 
-export const vars = {space};
+export const vars = {space, spaceMobile, spaceTablet, spaceDesktop};
+
+export const marginInline = style({
+    marginTop: `calc(${space} * -1)`,
+    marginLeft: `calc(${space} * -1)`,
+    '@media': {
+        [mq.mobile]: {
+            vars: {
+                [space]: spaceMobile,
+            },
+        },
+        [mq.tablet]: {
+            vars: {
+                [space]: fallbackVar(spaceTablet, spaceMobile),
+            },
+        },
+        [mq.desktopOrBigger]: {
+            vars: {
+                [space]: spaceDesktop,
+            },
+        },
+    },
+});
+
+export const flexInline = style({
+    justifyContent: space,
+});
 
 /**
  * https://caniuse.com/flexbox-gap
@@ -57,9 +87,12 @@ export const noFullWidth = style([
     },
 ]);
 
-globalStyle(`${inline} > div`, {
+globalStyle(`${marginInline} > div`, {
     marginLeft: space,
     marginTop: space,
+});
+
+globalStyle(`${inline} > div`, {
     pointerEvents: 'auto', // restore pointer events for children
     // Hack to fix https://jira.tid.es/browse/WEB-1683
     // In iOS the inline component sometimes cuts the last line of the content
@@ -84,16 +117,4 @@ globalStyle(`${inline} > div:first-child`, {
 
 globalStyle(`${inline} > div:empty`, {
     display: 'none',
-});
-
-export const justifyVariants = styleVariants({
-    between: {
-        justifyContent: 'space-between',
-    },
-    around: {
-        justifyContent: 'space-around',
-    },
-    evenly: {
-        justifyContent: 'space-evenly',
-    },
 });